+++ src/id_translation/_config_utils.py
@@ -2,13 +2,14 @@ import json
 import logging
 import sys
 from dataclasses import dataclass
+from hashlib import sha256
 from pathlib import Path
 from typing import TYPE_CHECKING, Any, List, Tuple, Type
 
 if sys.version_info >= (3, 11):
     import tomllib
 else:
-    import tomli as tomllib
+    import tomli as tomllib  # pragma: no cover
 
 import pandas as pd
 
@@ -28,10 +29,10 @@ class ConfigMetadata:
     """The ``rics`` version under which this instance was created."""
     created: pd.Timestamp
     """The time at which the ``Translator`` was originally initialized. Second precision."""
-    path: Path
-    """Absolute path of the main translation configuration."""
-    extra_fetchers: Tuple[Path, ...]
-    """Absolute paths of configuration files for auxiliary fetchers."""
+    main: Tuple[Path, str]
+    """Absolute path and fingerprint of the main translation configuration."""
+    extra_fetchers: Tuple[Tuple[Path, str], ...]
+    """Absolute path and fingerprint of configuration files for auxiliary fetchers."""
     clazz: str
     """String representation of the class type."""
 
@@ -42,7 +43,7 @@ class ConfigMetadata:
 
             - They have the same ``rics`` version, and
             - Use the same fully qualified class name, and
-            - The main configuration files are equal after parsing, and
+            - The main configuration is equal after parsing, and
             - They have the same number of auxiliary (`"extra"`) fetcher configurations, and
             - All auxiliary fetcher configurations are equal after parsing.
 
@@ -66,19 +67,25 @@ class ConfigMetadata:
             LOGGER.debug(f"Class not equal. Expected '{self.clazz}', but got '{other.clazz}'.")
             return False
 
-        if tomllib.loads(self.path.read_text()) != tomllib.loads(other.path.read_text()):
+        if self.main[1] != other.main[1]:
+            LOGGER.debug(f"Main configuration changed. Expected fingerprint {self.main[1]}, but got {other.main[1]}.")
             return False
 
         if len(self.extra_fetchers) != len(other.extra_fetchers):
             LOGGER.debug(
                 f"Number of auxiliary fetchers changed. Expected {len(self.extra_fetchers)}"
-                f" but got {len(other.extra_fetchers)}."
+                f" fetchers but got {len(other.extra_fetchers)}."
             )
             return False
 
         def func(i: int) -> bool:
-            if tomllib.loads(self.extra_fetchers[i].read_text()) != tomllib.loads(other.extra_fetchers[i].read_text()):
-                LOGGER.debug(f"Configuration has changed for auxiliary fetcher at {self.extra_fetchers[i]}.")
+            path, fingerprint = self.extra_fetchers[i]
+            _, other_fingerprint = other.extra_fetchers[i]
+            if fingerprint != other_fingerprint:
+                LOGGER.debug(
+                    f"Configuration changed for auxiliary fetcher #{i} at {path}. "
+                    f"Expected fingerprint {fingerprint}, but got {other_fingerprint}."
+                )
                 return False
             return True
 
@@ -91,10 +98,10 @@ class ConfigMetadata:
             id_translation_version=raw.pop("id_translation_version"),
             rics_version=raw.pop("rics_version"),
             created=raw.pop("created").isoformat(),
-            path=str(raw.pop("path")),
-            extra_fetchers=list(map(str, raw.pop("extra_fetchers"))),
-            clazz=raw.pop("clazz"),
+            main=tuple(map(str, raw.pop("main"))),
+            extra_fetchers=[tuple(map(str, t)) for t in raw.pop("extra_fetchers")],
         )
+        kwargs["class"] = raw.pop("clazz")
         assert not raw, f"Not serialized: {raw}."  # noqa:  S101
         return json.dumps(kwargs, indent=True)
 
@@ -102,13 +109,17 @@ class ConfigMetadata:
     def from_json(cls, s: str) -> "ConfigMetadata":
         """Create ``ConfigMetadata`` from a JSON string `s`."""
         raw = json.loads(s)
+
+        def to_path_tuple(args: List[str]) -> Tuple[Path, str]:
+            return Path(args[0]), args[1]
+
         kwargs = dict(
             id_translation_version=raw.pop("id_translation_version"),
             rics_version=raw.pop("rics_version"),
             created=pd.Timestamp.fromisoformat(raw.pop("created")),
-            path=Path(raw.pop("path")),
-            extra_fetchers=tuple(map(Path, raw.pop("extra_fetchers"))),
-            clazz=raw.pop("clazz"),
+            main=to_path_tuple(raw.pop("main")),
+            extra_fetchers=tuple(map(to_path_tuple, raw.pop("extra_fetchers"))),
+            clazz=raw.pop("class"),
         )
         assert not raw, f"Not deserialized: {raw}."  # noqa:  S101
         return ConfigMetadata(**kwargs)
@@ -124,22 +135,25 @@ def make_metadata(
 
     from id_translation import __version__ as id_translation_version
 
-    def fully_qualified_name(t: Type["Translator[Any, Any, Any]"]) -> str:
-        return t.__module__ + "." + t.__qualname__
+    def create_path_tuple(str_path: str) -> Tuple[Path, str]:
+        p = Path(str_path).expanduser().absolute()
+        with p.open("rb") as f:
+            content = tomllib.load(f)
+        return p, sha256(json.dumps(content).encode()).hexdigest()
 
     return ConfigMetadata(
         id_translation_version=id_translation_version,
         rics_version=rics_version,
         created=pd.Timestamp.now().round("s"),
-        path=Path(path).absolute(),
-        extra_fetchers=tuple(map(Path.absolute, map(Path, extra_fetchers))),
-        clazz=fully_qualified_name(clazz),
+        main=create_path_tuple(path),
+        extra_fetchers=tuple(map(create_path_tuple, extra_fetchers)),
+        clazz=clazz.__module__ + "." + clazz.__qualname__,  # Fully qualified name
     )
 
 
 def use_cached_translator(
     metadata_path: Path,
-    reference_metadata: ConfigMetadata,
+    wanted_config: ConfigMetadata,
     max_age: pd.Timedelta,
 ) -> bool:
     """Returns ``True`` if given metadata indicates that the cached ``Translator`` is still viable."""
@@ -147,20 +161,19 @@ def use_cached_translator(
         LOGGER.info(f"Metadata file '{metadata_path}' does not exist. Create new Translator.")
         return False
 
-    now = pd.Timestamp.now().round("s")
+    stored_config = ConfigMetadata.from_json(metadata_path.read_text())
+    LOGGER.debug(f"Metadata found: {stored_config}")
 
-    metadata = ConfigMetadata.from_json(metadata_path.read_text())
-    LOGGER.debug(f"Metadata found: {metadata}")
-
-    if not reference_metadata.is_equivalent(metadata):
+    if not wanted_config.is_equivalent(stored_config):
+        LOGGER.info(f"Configuration has changed. Reject cached Translator in '{metadata_path.parent}'.")
         return False
 
-    expires_at = metadata.created + max_age
-    age = abs(now - expires_at)
+    expires_at = stored_config.created + max_age
+    offset = abs(pd.Timestamp.now() - expires_at).round("s")
 
-    if expires_at < metadata.created:
-        LOGGER.info(f"Reject cached Translator in '{metadata_path.parent}'. Expired at {expires_at} ({age} ago).")
+    if expires_at <= stored_config.created:
+        LOGGER.info(f"Reject cached Translator in '{metadata_path.parent}'. Expired at {expires_at} ({offset} ago).")
         return False
 
-    LOGGER.info(f"Accept cached Translator in '{metadata_path.parent}'. Expires at {expires_at} (in {age}).")
+    LOGGER.info(f"Accept cached Translator in '{metadata_path.parent}'. Expires at {expires_at} (in {offset}).")
     return True